Improved SQL and removed redundant forum class
The forum class only served to hold an ID for each game's forum when the forum page could just be generated from selecting all threads in the database that are linked to the game that the forum entity would originally link to
